chore: remove theme context
- Removed theme context as everything is handled by tailwind classes
