+++ gatsby-config.js
@@ -49,7 +49,6 @@ module.exports = {
         alias: {
           "@components": path.resolve(__dirname, "./src/components"),
           "@utils": path.resolve(__dirname, "./src/utils"),
-          "@context": path.resolve(__dirname, "./src/context"),
         },
       },
     },
