+++ src/components/layout.tsx
@@ -1,23 +1,35 @@
-import React from "react"
+import React, { useEffect } from "react"
 
-import { ThemeProvider } from "../context/ThemeContext"
 import { LayoutProps } from "../types"
 import Footer from "./footer"
 import Header from "./header"
 import ScrollToTop from "./ScrollToTop"
 
 const Layout: React.FC<LayoutProps> = ({ children }) => {
+  useEffect(() => {
+    const mediaQuery = window.matchMedia("(prefers-color-scheme: dark)")
+
+    const handleChange = (e: MediaQueryListEvent) => {
+      const newTheme = e.matches ? "dark" : "light"
+      document.documentElement.setAttribute("data-theme", newTheme)
+    }
+
+    mediaQuery.addEventListener("change", handleChange)
+
+    return () => {
+      mediaQuery.removeEventListener("change", handleChange)
+    }
+  }, [])
+
   return (
-    <ThemeProvider>
-      <div className="m-0 flex min-h-screen w-full flex-col overflow-x-hidden">
-        <Header />
-        <main className="mx-auto w-full max-w-4xl flex-grow px-3 sm:px-4 lg:px-6">
-          {children}
-        </main>
-        <Footer />
-        <ScrollToTop />
-      </div>
-    </ThemeProvider>
+    <div className="m-0 flex min-h-screen w-full flex-col overflow-x-hidden">
+      <Header />
+      <main className="mx-auto w-full max-w-4xl flex-grow px-3 sm:px-4 lg:px-6">
+        {children}
+      </main>
+      <Footer />
+      <ScrollToTop />
+    </div>
   )
 }
