--- src/context/ThemeContext.tsx
@@ -1,81 +1,0 @@
-import React, {
-  createContext,
-  ReactNode,
-  useContext,
-  useEffect,
-  useState,
-} from "react"
-
-export type Theme = "light" | "dark"
-
-interface ThemeContextType {
-  theme: Theme
-  setTheme: (theme: Theme) => void
-}
-
-const ThemeContext = createContext<ThemeContextType | undefined>(undefined)
-
-export const useTheme = (): ThemeContextType => {
-  const context = useContext(ThemeContext)
-  if (!context) {
-    throw new Error("useTheme must be used within a ThemeProvider")
-  }
-  return context
-}
-
-interface ThemeProviderProps {
-  children: ReactNode
-}
-
-export const ThemeProvider: React.FC<ThemeProviderProps> = ({ children }) => {
-  // Get initial theme from the already-set attribute (from SSR script)
-  const getInitialTheme = (): Theme => {
-    if (typeof window !== "undefined") {
-      const currentTheme = document.documentElement.getAttribute("data-theme")
-      if (currentTheme === "dark" || currentTheme === "light") {
-        return currentTheme
-      }
-      return window.matchMedia("(prefers-color-scheme: dark)").matches
-        ? "dark"
-        : "light"
-    }
-    return "light"
-  }
-
-  const [theme, setThemeState] = useState<Theme>(getInitialTheme)
-
-  useEffect(() => {
-    // Sync with current theme (already set by SSR script, so no flash)
-    const currentTheme = document.documentElement.getAttribute(
-      "data-theme"
-    ) as Theme
-    if (currentTheme && currentTheme !== theme) {
-      setThemeState(currentTheme)
-    }
-
-    // Listen for system theme changes
-    const mediaQuery = window.matchMedia("(prefers-color-scheme: dark)")
-    const handleChange = (e: MediaQueryListEvent) => {
-      const newTheme = e.matches ? "dark" : "light"
-      setThemeState(newTheme)
-      document.documentElement.setAttribute("data-theme", newTheme)
-    }
-
-    mediaQuery.addEventListener("change", handleChange)
-
-    return () => {
-      mediaQuery.removeEventListener("change", handleChange)
-    }
-  }, [])
-
-  const setTheme = (newTheme: Theme) => {
-    setThemeState(newTheme)
-    document.documentElement.setAttribute("data-theme", newTheme)
-  }
-
-  return (
-    <ThemeContext.Provider value={{ theme, setTheme }}>
-      {children}
-    </ThemeContext.Provider>
-  )
-}
