+++ tsconfig.json
@@ -19,7 +19,6 @@
     "paths": {
       "@components/*": ["src/components/*"],
       "@utils/*": ["src/utils/*"],
-      "@context/*": ["src/context/*"],
       "@styles/*": ["src/styles/*"]
     }
   },
